welele

+++ src/BLL.Services/Extensions/TicketExtensions.cs
@@ -8,7 +8,7 @@
     {
         public static decimal GetScore(this IEnumerable<Ticket> tickets)
         {
-            return tickets.Where(t=>t.Priority>0).Sum(t => 1/t.Priority);
+            return tickets.Where(t=>t.Priority>0).Sum(t => 1/ (decimal)t.Priority);
         }
         public static IEnumerable<string> GetChanges(this Ticket ticket, Ticket otherTicket)
         {
